Removed two phase bandpass filter instead added two phase baselines. Changed curve fit to use medians instead of IIR lowpass filter. Changed stats for quality to use only 1 second instead of 2.5 from history.

+++ app/src/main/java/care/dovetail/tracker/Config.java
@@ -5,7 +5,7 @@ public class Config {
 
     public static final int BLUETOOTH_ENABLE_REQUEST = 0;
 
-    public static final double SAMPLING_FREQ  = 200.0;
+    public static final int SAMPLING_FREQ  = 200;
 
     public static final int BLINK_WINDOW  = 20;
 }
